fix: improve ConsoleMessage text() results
