+++ packages/puppeteer-core/src/cdp/ExecutionContext.ts
@@ -36,7 +36,7 @@ import {
   addPageBinding,
   CDP_BINDING_PREFIX,
   createEvaluationError,
-  valueFromRemoteObject,
+  valueFromPrimitiveRemoteObject,
 } from './utils.js';
 
 const ariaQuerySelectorBinding = new Binding(
@@ -403,7 +403,7 @@ export class ExecutionContext
       }
 
       if (returnByValue) {
-        return valueFromRemoteObject(remoteObject) as HandleFor<
+        return valueFromPrimitiveRemoteObject(remoteObject) as HandleFor<
           Awaited<ReturnType<Func>>
         >;
       }
@@ -457,9 +457,9 @@ export class ExecutionContext
     }
 
     if (returnByValue) {
-      return valueFromRemoteObject(remoteObject) as unknown as HandleFor<
-        Awaited<ReturnType<Func>>
-      >;
+      return valueFromPrimitiveRemoteObject(
+        remoteObject,
+      ) as unknown as HandleFor<Awaited<ReturnType<Func>>>;
     }
 
     return this.#world.createCdpHandle(remoteObject) as HandleFor<
